factored out dispatch table building into separate header file
